+++ R/function_PlotBasinSummary.R
@@ -1,7 +1,7 @@
 #' Plot a summary of model results for a single sub-basin
 #' 
 #' Plot a standard suite of plots summarizing properties of a sub-basin including upstream area and model performance 
-#' for discharge and concentrations of nutrients, sediment, and tracers.
+#' for discharge and concentrations of nutrients, organic carbon, sediment, and tracers.
 #' 
 #' @param x Data frame, with column-wise daily time series of HYPE variables. Date-times in 
 #' \code{\link{POSIXct}} format in first column. Typically an imported basin output file from HYPE using \code{\link{ReadBasinOutput}}. 
@@ -44,7 +44,7 @@
 #' @details
 #' \code{PlotBasinSummary} plots a multi-panel plot with a number of plots to evaluate model properties and performances for a 
 #' chosen sub-basin. Performance plots include discharge, HYPE-modeled nutrient species for nitrogen (total, inorganic, organic) 
-#' and phosphorus (total, particulate, soluble), and HYPE modeled suspended and total sediment concentrations.
+#' phosphorus (total, particulate, soluble), organic carbon, and suspended and total sediment concentrations.
 #' 
 #' Plotted panels show: 
 #' \itemize{
@@ -54,7 +54,7 @@
 #' \item{\emph{Goodness-of-fit measures for discharge and concentrations}: KGE (Kling-Gupta Efficiency), NSE (Nash-Sutcliffe Efficiency), PBIAS 
 #' (Percentage Bias, aka relative error), MAE (Mean Absolute Error), r (Pearson product-moment correlation coefficient), VE (Volumetric Efficiency).}
 #' \item{\emph{Simulation-observation relationships for discharge and concentrations}: Simulated and observed concentration-discharge relationships, 
-#' relationship between observed and simulated nutrient, sediment, and tracer concentrations.}
+#' relationship between observed and simulated nutrient, organic carbon, sediment, and tracer concentrations.}
 #' \item{\emph{Duration curves for flow and concentrations}: Pairwise simulated and observed curves.}
 #' \item{\emph{Annual regimes for flow and concentrations}: Pairwise simulated and observed regime plots at monthly aggregation, with 
 #' number of observations for concentration regimes.}