+++ NEWS.md
@@ -5,6 +5,7 @@
 
 #### Highlights
 * Fix bug in AllUpstreamSubids() that resulted in incorrect upstream SUBIDs when MAINPART == 0 in BranchData.txt
+* Add plots for modelled organic carbon in PlotBasinOutput() and PlotBasinSummary()
 
 ## HYPEtools 1.6.4 (2024-11-27)
 *Enhancements and bug fixes*